Rename and Reorganise listing algorithms
