+++ lib/algos/list/MPU.js
@@ -9,9 +9,9 @@ function numberDefault(num, defaultNum) {
 }
 
 /**
- *  Class for the ListMultipartUploads extension
+ *  Class for the MultipartUploads extension
  */
-class ListMultipartUploads {
+class MultipartUploads {
     /**
      *  Constructor of the extension
      *  Init and check parameters
@@ -129,5 +129,5 @@ class ListMultipartUploads {
 }
 
 module.exports = {
-    ListMultipartUploads,
+    MultipartUploads,
 };
