+++ lib/algos/list/delimiter.js
