+++ lib/algos/list/tools.js
